+++ src/components/BackgroundPickerAndDarkModeToggle.tsx
@@ -15,6 +15,11 @@ export const BackgroundPickerAndDarkModeToggle = ({
 }) => (
   <div style={{ display: "flex" }}>
     {actionManager.renderAction("changeViewBackgroundColor")}
-    {showThemeBtn && <>{actionManager.renderAction("toggleTheme")}</>}
+    {showThemeBtn && actionManager.renderAction("toggleTheme")}
+    {appState.fileHandle && (
+      <div style={{ marginInlineStart: "0.25rem" }}>
+        {actionManager.renderAction("saveToActiveFile")}
+      </div>
+    )}
   </div>
 );